+++ lib/metro/constants/strings.dart
@@ -1,0 +1,14 @@
+// argument flags
+const String fileOption = 'file';
+const String helpFlag = 'help';
+const String storableFlag = 'storable';
+const String forceFlag = 'force';
+const String controllerFlag = 'controller';
+
+// folders
+const String yamlPath = 'pubspec.yaml';
+const String controllerFolder = 'lib/app/controllers/';
+const String widgetFolder = 'lib/resources/widgets/';
+const String pageFolder = 'lib/resources/pages/';
+const String modelFolder = 'lib/app/models/';
+const String networkServicesFolder = 'lib/app/networking/';