+++ lib/metro/metro.dart
@@ -1,77 +1,96 @@
 library metro;
 
-import 'dart:convert';
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter_launcher_icons/main.dart' as launcherIcons;
-import 'package:http/http.dart' as http;
 import 'package:args/args.dart';
-import 'package:json_to_dart/json_to_dart.dart';
-
-import 'package:nylo_framework/metro/constants/arg_flags.dart';
-import 'package:nylo_framework/metro/constants/folders.dart';
-
-import 'package:nylo_framework/metro/models/ny_api_request.dart';
-import 'package:nylo_framework/metro/networking/api_request_networking.dart';
-import 'package:nylo_framework/metro/stubs/api_request_stub.dart';
-import 'package:nylo_framework/metro/stubs/api_service_stub.dart';
+import 'package:nylo_framework/metro/constants/strings.dart';
+import 'package:nylo_framework/metro/models/ny_command.dart';
 import 'package:nylo_framework/metro/stubs/controller_stub.dart';
 import 'package:nylo_framework/metro/stubs/model_stub.dart';
 import 'package:nylo_framework/metro/stubs/page_stub.dart';
 import 'package:nylo_framework/metro/stubs/page_w_controller_stub.dart';
-import 'package:nylo_framework/metro/stubs/widget_stub.dart';
-
-import 'helpers/tools.dart';
-import 'models/ny_command.dart';
+import 'package:nylo_framework/metro/stubs/widget_stateful_stub.dart';
+import 'package:nylo_framework/metro/stubs/widget_stateless_stub.dart';
 import 'dart:io';
 
-List<NyCommand> _allMakeCommands = [
-  NyCommand(name: "controller", options: 1, arguments: ["-h", "-f"], action: _makeController),
-  NyCommand(name: "model", options: 1, arguments: ["-h"], action: _makeModel),
-  NyCommand(name: "page", options: 1, arguments: ["-h"], action: _makePage),
-  NyCommand(name: "widget", options: 1, arguments: ["-h"], action: _makeWidget),
+List<NyCommand> _allProjectCommands = [
+  NyCommand(name: "init", options: 1, arguments: ["-h"], action: _projectInit),
 ];
 
-List<NyCommand> _allApiSpecCommands = [
-  NyCommand(name: "build", options: 1, arguments: ["-h"], action: _apiSpecBuild),
+List<NyCommand> _allMakeCommands = [
+  NyCommand(
+      name: "controller",
+      options: 1,
+      arguments: ["-h", "-f"],
+      action: _makeController),
+  NyCommand(
+      name: "model",
+      options: 1,
+      arguments: ["-h", "-f", "-s"],
+      action: _makeModel),
+  NyCommand(
+      name: "page", options: 1, arguments: ["-h", "-f"], action: _makePage),
+  NyCommand(
+      name: "stateful_widget",
+      options: 1,
+      arguments: ["-h", "-f"],
+      action: _makeStatefulWidget),
+  NyCommand(
+      name: "stateless_widget",
+      options: 1,
+      arguments: ["-h", "-f"],
+      action: _makeStatelessWidget),
 ];
 
 List<NyCommand> _allAppIconsCommands = [
-  NyCommand(name: "build", options: 1, arguments: ["-h"], action: _makeAppIcons),
+  NyCommand(
+      name: "build", options: 1, arguments: ["-h"], action: _makeAppIcons),
 ];
 
 Future<void> commands(List<String> arguments) async {
-
   if (arguments.length == 0) {
     _handleMenu();
-    return;
+    exit(0);
   }
 
   List<String> argumentSplit = arguments[0].split(":");
 
-  if (argumentSplit.length == 0) {
+  if (argumentSplit.length == 0 || argumentSplit.length <= 1) {
     writeInBlack('Invalid arguments ' + arguments.toString());
-    return;
+    exit(2);
   }
 
   String type = argumentSplit[0];
   String action = argumentSplit[1];
 
-  NyCommand nyCommand;
+  NyCommand? nyCommand;
   switch (type) {
-    case "make": {
-      nyCommand = _allMakeCommands.firstWhere((command) => command.name == action, orElse: () => null);
-      break;
-    }
-    case "apispec": {
-      nyCommand = _allApiSpecCommands.firstWhere((command) => command.name == action, orElse: () => null);
-      break;
-    }
-    case "appicons": {
-      nyCommand = _allAppIconsCommands.firstWhere((command) => command.name == action, orElse: () => null);
-      break;
-    }
-    default: {
-
-    }
+    case "project":
+      {
+        nyCommand = _allProjectCommands
+            .firstWhereOrNull((command) => command.name == action);
+        break;
+      }
+    case "make":
+      {
+        nyCommand = _allMakeCommands
+            .firstWhereOrNull((command) => command.name == action);
+        break;
+      }
+    case "appicons":
+      {
+        nyCommand = _allAppIconsCommands
+            .firstWhereOrNull((command) => command.name == action);
+        break;
+      }
+    case "plugin":
+      {
+        nyCommand = _allAppIconsCommands
+            .firstWhereOrNull((command) => command.name == action);
+        break;
+      }
+    default:
+      {}
   }
 
   if (nyCommand == null) {
@@ -80,11 +99,12 @@ Future<void> commands(List<String> arguments) async {
   }
 
   arguments.removeAt(0);
-  nyCommand.action(arguments);
+  nyCommand.action!(arguments);
 }
 
 _handleMenu() {
-  writeInBlack('Metro - Nylo\'s Companion to Build Flutter apps by Anthony Gordon');
+  writeInBlack(
+      'Metro - Nylo\'s Companion to Build Flutter apps by Anthony Gordon');
   writeInBlack('Version: 1.0.0');
 
   writeInBlack('');
@@ -100,56 +120,131 @@ _handleMenu() {
 
   writeInBlack('All commands:');
 
-  writeInGreen(' make');
+  _writeInGreen(' project');
+  _allProjectCommands.forEach((command) {
+    writeInBlack('  project:' + command.name!);
+  });
+
+  writeInBlack('');
+
+  _writeInGreen(' make');
   _allMakeCommands.forEach((command) {
-    writeInBlack('  make:' + command.name);
+    writeInBlack('  make:' + command.name!);
   });
 
   writeInBlack('');
 
-  writeInGreen(' appicons');
-  _allApiSpecCommands.forEach((command) {
-    writeInBlack('  appicons:' + command.name);
+  _writeInGreen(' appicons');
+  _allAppIconsCommands.forEach((command) {
+    writeInBlack('  appicons:' + command.name!);
   });
-
-  writeInBlack('');
-
-  writeInGreen(' apispec');
-  _allApiSpecCommands.forEach((command) {
-    writeInBlack('  apispec:' + command.name);
-  });
-
 }
 
-_makeWidget(List<String> arguments) async {
+_projectInit(List<String> arguments) async {
   final ArgParser parser = ArgParser(allowTrailingOptions: true);
 
-  parser.addFlag(helpFlag, abbr: 'h', help: 'e.g. make:controller DashboardController', negatable: false);
+  parser.addFlag(helpFlag,
+      abbr: 'h', help: 'Initializes the Nylo project', negatable: false);
 
   final ArgResults argResults = parser.parse(arguments);
 
   if (argResults[helpFlag]) {
-    writeInBlack('Help - Creates a new controller in Nylo');
+    writeInBlack('Help - Initializes the Nylo project');
     writeInBlack(parser.usage);
     exit(0);
   }
 
-  List<String> split = argResults.arguments.first.split("_");
-  List<String> map = split.map((e) => capitalize(e)).toList();
-  String newStr = map.join("");
+  final File envExample = File('.env-example');
+  final File env = File('.env');
 
-  String path = '$widgetFolder/${argResults.arguments.first}.dart';
+  await env.writeAsString(await envExample.readAsString());
 
-  if (await File(path).exists()) {
-    writeInRed(argResults.arguments.first + ' already exists');
-    return;
+  _writeInGreen('Project initialized, create something great 🎉');
+  exit(0);
+}
+
+_makeStatefulWidget(List<String> arguments) async {
+  final ArgParser parser = ArgParser(allowTrailingOptions: true);
+
+  parser.addFlag(helpFlag,
+      abbr: 'h',
+      help: 'e.g. make:widget video_player_widget',
+      negatable: false);
+  parser.addFlag(forceFlag,
+      abbr: 'f',
+      help: 'Creates a new widget even if it already exists.',
+      negatable: false);
+
+  final ArgResults argResults = parser.parse(arguments);
+
+  bool? hasForceFlag = argResults[forceFlag];
+  bool hasHelpFlag = argResults[helpFlag];
+
+  if (hasHelpFlag) {
+    writeInBlack(parser.usage);
+    exit(0);
+  }
+
+  String widgetName =
+      argResults.arguments.first.replaceAll(RegExp(r'(_?widget)'), "");
+
+  String path = '$widgetFolder/${widgetName.toLowerCase()}_widget.dart';
+
+  if (await File(path).exists() && hasForceFlag == false) {
+    _writeInRed(widgetName + '_widget already exists');
+    exit(0);
   }
 
   final File file = File(path);
 
-  await file.writeAsString(widgetStub(newStr));
+  await _makeDirectory(widgetFolder);
 
-  writeInGreen(argResults.arguments.first + ' created 🎉');
+  await file.writeAsString(widgetStatefulStub(_parseToPascal(widgetName)));
+
+  _writeInGreen(widgetName + '_widget created 🎉');
+
+  exit(0);
+}
+
+_makeStatelessWidget(List<String> arguments) async {
+  final ArgParser parser = ArgParser(allowTrailingOptions: true);
+
+  parser.addFlag(helpFlag,
+      abbr: 'h',
+      help: 'e.g. make:widget video_player_widget',
+      negatable: false);
+  parser.addFlag(forceFlag,
+      abbr: 'f',
+      help: 'Creates a new widget even if it already exists.',
+      negatable: false);
+
+  final ArgResults argResults = parser.parse(arguments);
+
+  bool? hasForceFlag = argResults[forceFlag];
+  bool hasHelpFlag = argResults[helpFlag];
+
+  if (hasHelpFlag) {
+    writeInBlack(parser.usage);
+    exit(0);
+  }
+
+  String widgetName =
+      argResults.arguments.first.replaceAll(RegExp(r'(_?widget)'), "");
+
+  String path = '$widgetFolder/${widgetName.toLowerCase()}_widget.dart';
+
+  if (await File(path).exists() && hasForceFlag == false) {
+    _writeInRed(widgetName + '_widget already exists');
+    exit(0);
+  }
+
+  final File file = File(path);
+
+  await _makeDirectory(widgetFolder);
+
+  await file.writeAsString(widgetStatelessStub(_parseToPascal(widgetName)));
+
+  _writeInGreen(widgetName + '_widget created 🎉');
 
   exit(0);
 }
@@ -157,7 +252,10 @@ _makeWidget(List<String> arguments) async {
 _makeAppIcons(List<String> arguments) async {
   final ArgParser parser = ArgParser(allowTrailingOptions: true);
 
-  parser.addFlag(helpFlag, abbr: 'h', help: 'e.g. make:app_icons', negatable: false);
+  parser.addFlag(helpFlag,
+      abbr: 'h',
+      help: 'Generates your app icons in the project.',
+      negatable: false);
 
   final ArgResults argResults = parser.parse(arguments);
 
@@ -169,16 +267,21 @@ _makeAppIcons(List<String> arguments) async {
 
   launcherIcons.createIconsFromArguments(arguments);
 
-  writeInGreen('App icons created 🎉');
+  _writeInGreen('App icons created 🎉');
   exit(0);
 }
-
 
 _makeController(List<String> arguments) async {
   final ArgParser parser = ArgParser(allowTrailingOptions: true);
 
-  parser.addFlag(helpFlag, abbr: 'h', help: 'Used to make new controllers e.g. home_controller', negatable: false);
-  parser.addFlag(forceFlag, abbr: 'f', help: 'Creates a new controller even if it already exists.', negatable: false);
+  parser.addFlag(helpFlag,
+      abbr: 'h',
+      help: 'Used to make new controllers e.g. home_controller',
+      negatable: false);
+  parser.addFlag(forceFlag,
+      abbr: 'f',
+      help: 'Creates a new controller even if it already exists.',
+      negatable: false);
 
   final ArgResults argResults = parser.parse(arguments);
 
@@ -187,7 +290,7 @@ _makeController(List<String> arguments) async {
     exit(0);
   }
 
-  bool hasForceFlag = argResults[forceFlag];
+  bool? hasForceFlag = argResults[forceFlag];
   bool hasHelpFlag = argResults[helpFlag];
 
   if (hasHelpFlag) {
@@ -195,41 +298,38 @@ _makeController(List<String> arguments) async {
     exit(0);
   }
 
-  String firstArg = argResults.arguments.first;
-  String path = '$controllerFolder/$firstArg.dart';
+  String className =
+      argResults.arguments.first.replaceAll(RegExp(r'(_?controller)'), "");
+  String path = '$controllerFolder/${className.toLowerCase()}_controller.dart';
 
   if (await File(path).exists() && hasForceFlag == false) {
-    writeInRed(firstArg + ' already exists');
+    _writeInRed(className + ' already exists');
     exit(0);
   }
 
-  String controllerName = _parseToPascal(firstArg);
+  String controllerName = _parseToPascal(className);
   String strController = controllerStub(controllerName: controllerName);
 
   await _writeToFilePath(path, strController);
 
-  writeInGreen(firstArg + ' created 🎉');
+  _writeInGreen(className + '_controller created 🎉');
 
   exit(0);
-}
-
-String _parseToPascal(name) {
-  List<String> split = name.split("_");
-  List<String> map = split.map((e) => capitalize(e)).toList();
-  return map.join("");
-}
-
-_writeToFilePath(path, strFile) async {
-  final File file = File(path);
-
-  await file.writeAsString(strFile);
 }
 
 _makeModel(List<String> arguments) async {
   final ArgParser parser = ArgParser(allowTrailingOptions: true);
 
-  parser.addFlag(helpFlag, abbr: 'h', help: 'Used to make new controllers e.g. home_controller', negatable: false);
-  parser.addFlag(forceFlag, abbr: 'f', help: 'Creates a new controller even if it already exists.', negatable: false);
+  parser.addFlag(helpFlag,
+      abbr: 'h',
+      help: 'Creates a new model in your project.',
+      negatable: false);
+  parser.addFlag(forceFlag,
+      abbr: 'f',
+      help: 'Creates a new model even if it already exists.',
+      negatable: false);
+  parser.addFlag(storableFlag,
+      abbr: 's', help: 'Create a new Storable model.', negatable: false);
 
   final ArgResults argResults = parser.parse(arguments);
 
@@ -238,8 +338,9 @@ _makeModel(List<String> arguments) async {
     exit(0);
   }
 
-  bool hasForceFlag = argResults[forceFlag];
+  bool? hasForceFlag = argResults[forceFlag];
   bool hasHelpFlag = argResults[helpFlag];
+  bool? hasStorableFlag = argResults[storableFlag];
 
   if (hasHelpFlag) {
     writeInBlack(parser.usage);
@@ -247,21 +348,24 @@ _makeModel(List<String> arguments) async {
   }
 
   String firstArg = argResults.arguments.first;
-  String path = '$modelFolder/$firstArg.dart';
+  String path = '$modelFolder/${firstArg.toLowerCase()}.dart';
 
   if (await File(path).exists() && hasForceFlag == false) {
-    writeInRed(firstArg + ' already exists');
+    _writeInRed(firstArg + ' already exists');
     exit(0);
   }
 
   final File file = File(path);
 
   String modelName = _parseToPascal(firstArg);
-  String strModel = modelStub(modelName: modelName);
+  String strModel =
+      modelStub(modelName: modelName, isStorable: hasStorableFlag);
+
+  await _makeDirectory(modelFolder);
 
   await file.writeAsString(strModel);
 
-  writeInGreen(modelName + ' created 🎉');
+  _writeInGreen(modelName + ' created 🎉');
 
   exit(0);
 }
@@ -269,10 +373,18 @@ _makeModel(List<String> arguments) async {
 _makePage(List<String> arguments) async {
   final ArgParser parser = ArgParser(allowTrailingOptions: true);
 
-  parser.addFlag(helpFlag, abbr: 'h', help: 'e.g. make:page Menu', negatable: false);
+  parser.addFlag(helpFlag,
+      abbr: 'h',
+      help: 'Creates a new page widget for your project.',
+      negatable: false,
+  );
 
   bool shouldCreateController = false;
-  parser.addFlag("controller", abbr: 'c', help: 'Creates a new page with a controller', negatable: false);
+  parser.addFlag(controllerFlag,
+      abbr: 'c',
+      help: 'Creates a new page with a controller',
+      negatable: false,
+  );
 
   final ArgResults argResults = parser.parse(arguments);
 
@@ -290,136 +402,92 @@ _makePage(List<String> arguments) async {
     exit(0);
   }
 
-  String optionName;
+  String className =
+      argResults.arguments.first.replaceAll(RegExp(r'(_?page)'), "");
 
-  optionName = capitalize(argResults.arguments.first.replaceAll("_page", ""));
-
-  List<String> split = optionName.split("_");
-  List<String> map = split.map((e) => capitalize(e)).toList();
-  String className = map.join("");
-
-  String pathPage = '$pageFolder/${optionName.toLowerCase()}_page.dart';
-  String pathController = '$controllerFolder/${optionName.toLowerCase()}_controller.dart';
+  String pathPage = '$pageFolder/${className.toLowerCase()}_page.dart';
+  String pathController =
+      '$controllerFolder/${className.toLowerCase()}_controller.dart';
   if (shouldCreateController) {
     if (await File(pathPage).exists()) {
-      writeInRed(optionName + 'Page already exists');
+      _writeInRed('${className}_page already exists');
       return;
     }
 
     if (await File(pathController).exists()) {
-      writeInRed(optionName + 'Controller already exists');
+      _writeInRed('${className}_controller already exists');
       return;
     }
 
     final File filePage = File(pathPage);
-    await filePage.writeAsString(pageWithControllerStub(className: className, importName: argResults.arguments.first.replaceAll("_page", "")));
 
-    final File fileInterface = File(pathController);
+    await _makeDirectory(controllerFolder);
 
-    String strInterface = controllerStub(controllerName: className);
+    await filePage.writeAsString(pageWithControllerStub(
+        className: _parseToPascal(className),
+        importName: argResults.arguments.first.replaceAll("_page", "")));
 
-    await fileInterface.writeAsString(strInterface);
+    final File fileController = File(pathController);
 
-    writeInGreen('${className}Page & ${className}Controller created 🎉');
+    String strController =
+        controllerStub(controllerName: _parseToPascal(className));
+
+    await fileController.writeAsString(strController);
+
+    _writeInGreen('${className}_page & ${className}_controller created 🎉');
     exit(0);
   }
 
   if (await File(pathPage).exists()) {
-    writeInRed(argResults.arguments.first + ' already exists');
+    _writeInRed(argResults.arguments.first + ' already exists');
     return;
   }
 
   final File file = File(pathPage);
 
-  await file.writeAsString(pageStub(pageName: className));
+  await _makeDirectory(pageFolder);
 
-  writeInGreen('${optionName}Page created 🎉');
+  await file.writeAsString(pageStub(pageName: _parseToPascal(className)));
+
+  _writeInGreen('${className}_page created 🎉');
 
   exit(0);
 }
 
-_apiSpecBuild(List<String> arguments) async {
-  writeInBlack("Building API Spec\n");
-
-  File file;
-  Iterable json;
-
-  try {
-
-    file = File("apispec.json");
-    json = jsonDecode(await file.readAsString());
-
-  } on Exception catch (e) {
-    writeInRed("Error");
-    writeInBlack("Please check your apispec.yaml");
-    writeInBlack(e.toString());
-    return;
-  }
-
-  List<NyApiRequest> nyApiRequests = List.of(json).map((e) => NyApiRequest.fromJson(e)).toList();
-
-  String apiRequests = "";
-  String importApiRequests = "";
-
-  final ArgParser parser = ArgParser(allowTrailingOptions: true);
-
-  parser.addFlag(helpFlag, abbr: 'h', help: 'Used to auto build models from your apispec.yaml', negatable: false);
-  parser.addFlag(forceFlag, abbr: 'f', help: 'Forcefully creates models even if they exist.', negatable: false);
-
-  final ArgResults argResults = parser.parse(arguments);
-
-  bool hasForceFlag = argResults[forceFlag];
-  bool hasHelpFlag = argResults[helpFlag];
-
-  if (hasHelpFlag) {
-    writeInBlack(parser.usage);
-    exit(0);
-  }
-
-  for (int i = 0; i < nyApiRequests.length; i++) {
-
-    String modelPath = '$modelFolder/${nyApiRequests[i].modelName.toLowerCase()}.dart';
-
-    if (await File(modelPath).exists() && hasForceFlag == false) {
-      writeInBlue(nyApiRequests[i].modelName + ' already exists, skipping...');
-      continue;
-    }
-
-    http.Response json = await metroApiRequest(nyApiRequests[i]);
-
-    final classGenerator = new ModelGenerator(nyApiRequests[i].modelName);
-
-    DartCode dartCode = classGenerator.generateDartClasses(json.body);
-
-    final File file = File(modelPath);
-    await file.writeAsString(dartCode.code);
-    writeInBlue(nyApiRequests[i].modelName + " model created");
-
-    apiRequests = apiRequestStub(nyApiRequests[i]) + "\n" + apiRequests;
-
-    importApiRequests = "import \"../models/${nyApiRequests[i].modelName.toLowerCase()}.dart\";\n" + importApiRequests;
-  }
-
-  String networkingServicePath = '$networkServicesFolder/base_api_service.dart';
-  final File networkingServiceFile = File(networkingServicePath);
-  await networkingServiceFile.writeAsString(apiServiceStub(imports: importApiRequests, apiRequests: apiRequests));
-
-  writeInGreen("Api Spec Built 🚨");
-  exit(0);
-}
-
-writeInGreen(String message) {
+_writeInGreen(String message) {
   stdout.writeln('\x1B[92m' + message + '\x1B[0m');
 }
 
-writeInRed(String message) {
+_writeInRed(String message) {
   stdout.writeln('\x1B[91m' + message + '\x1B[0m');
-}
-
-writeInBlue(String message) {
-  stdout.writeln('\x1B[94m' + message + '\x1B[0m');
 }
 
 writeInBlack(String message) {
   stdout.writeln(message);
+}
+
+String _parseToPascal(name) {
+  List<String> split = name.split("_");
+  List<String> map = split.map((e) => capitalize(e)).toList();
+  return map.join("");
+}
+
+_writeToFilePath(path, strFile) async {
+  final File file = File(path);
+
+  await file.writeAsString(strFile);
+}
+
+_makeDirectory(String path) async {
+  final File dirFolder = File(path);
+  if (!(await dirFolder.exists())) {
+    await Directory(path).create();
+  }
+}
+
+String capitalize(String input) {
+  if (input.length == 0) {
+    return input;
+  }
+  return input[0].toUpperCase() + input.substring(1);
 }