+++ lib/metro/stubs/controller_stub.dart
@@ -1,12 +1,12 @@
-
-String controllerStub({String controllerName}) => '''
-import 'package:nylo_framework/metro/controllers/controller.dart';
-import 'package:flutter/material.dart';
+String controllerStub({String? controllerName}) => '''
+import 'controller.dart';
+import 'package:flutter/widgets.dart';
 
 class ${controllerName}Controller extends Controller {
   
-  ${controllerName}Controller.of(BuildContext context) {
-    super.context = context;
+  @override
+  construct(BuildContext context) {
+    
   }
 
 }