+++ lib/metro/stubs/page_stub.dart
@@ -1,6 +1,6 @@
-
-String pageStub({String pageName}) => '''
+String pageStub({String? pageName}) => '''
 import 'package:flutter/material.dart';
+import 'package:nylo_support/widgets/ny_state.dart';
 
 class ${pageName}Page extends StatefulWidget {
   
@@ -10,11 +10,11 @@ class ${pageName}Page extends StatefulWidget {
   _${pageName}PageState createState() => _${pageName}PageState();
 }
 
-class _${pageName}PageState extends State<${pageName}Page> {
+class _${pageName}PageState extends NyState<${pageName}Page> {
 
   @override
-  void initState() {
-    super.initState();
+  widgetDidLoad() async {
+  
   }
   
   @override
@@ -29,7 +29,7 @@ class _${pageName}PageState extends State<${pageName}Page> {
         
       ),
       body: SafeArea(
-        
+         child: Container(),
       ),
     );
   }