+++ lib/metro/stubs/page_w_controller_stub.dart
@@ -1,24 +1,27 @@
-
 import 'package:nylo_framework/metro/helpers/tools.dart';
 
-String pageWithControllerStub({String className, String importName}) => '''
+String pageWithControllerStub(
+        {String? className, required String importName}) =>
+    '''
 import 'package:flutter/material.dart';
+import 'package:nylo_support/widgets/ny_stateful_widget.dart';
 import '../../app/controllers/${importName.toLowerCase()}_controller.dart';
+import 'package:nylo_support/widgets/ny_state.dart';
 
-class ${capitalize(className)}Page extends StatefulWidget {
-  final ${capitalize(className)}Controller controller;
+class ${capitalize(className)}Page extends NyStatefulWidget {
+  final ${capitalize(className)}Controller controller = ${capitalize(className)}Controller();
   
-  ${capitalize(className)}Page({Key key, this.controller}) : super(key: key);
+  ${capitalize(className)}Page({Key key}) : super(key: key);
   
   @override
   _${capitalize(className)}PageState createState() => _${capitalize(className)}PageState();
 }
 
-class _${capitalize(className)}PageState extends State<${capitalize(className)}Page> {
+class _${capitalize(className)}PageState extends NyState<${capitalize(className)}Page> {
 
   @override
-  void initState() {
-    super.initState();
+  widgetDidLoad() async {
+  
   }
   
   @override
@@ -33,7 +36,7 @@ class _${capitalize(className)}PageState extends State<${capitalize(className)}P
         
       ),
       body: SafeArea(
-        
+        child: Container()
       ),
     );
   }